Agregando Registro y adaptando a todo el sistema

+++ Proyecto_11211070/object.cpp
@@ -6,7 +6,7 @@ Object::Object()
 
 string Object::toString() const {
     string retval = "Object@";
-    retval += this;
+    retval += (int) this;
     return retval;
 }
